# Optimize client with gRPC compression