Before:
```
                  Method |     Mean |     Error |    StdDev |     Op/s | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
------------------------ |---------:|----------:|----------:|---------:|------------:|------------:|------------:|--------------------:|
 CompressedSayHelloAsync | 12.61 us | 0.0985 us | 0.0873 us | 79,331.6 |      2.2278 |           - |           - |             9.16 KB |
```

After:
```
                  Method |     Mean |     Error |    StdDev |     Op/s | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
------------------------ |---------:|----------:|----------:|---------:|------------:|------------:|------------:|--------------------:|
 CompressedSayHelloAsync | 12.17 us | 0.1154 us | 0.1080 us | 82,173.4 |      2.1057 |           - |           - |             8.66 KB |
```